fix(layout): extract BottomTabBar route helpers to server-safe module (PR-BETA-6 hotfix #2)
Server Components in [locale]/layout.tsx cannot import from 'use client'
modules without crashing the render. Move BOTTOM_TAB_BAR_EXCLUDED_ROUTES,
stripLocalePrefix, isExcludedRoute to bottom-tab-bar.routes.ts so both
server and client can consume safely.

Fixes "Quelque chose s'est cassé" on all PR #182 preview Vercel pages.

- bottom-tab-bar.routes.ts: new server-safe module, no 'use client',
  no React/Next imports, pure data + pure functions.
- BottomTabBar.tsx: helpers removed from this 'use client' file, kept
  a JSDoc breadcrumb explaining the split.
- [locale]/layout.tsx: imports isExcludedRoute + stripLocalePrefix from
  the new module (BottomTabBar component import unchanged).
- BottomTabBar.test.tsx: helper imports point at the new module; the
  27 specs covering them still pass.

Build smoke: `npm run build` succeeds, all locale routes prerender as
dynamic ƒ (the bar mount lives behind a headers() read, expected).
Quality gates: lint 0 errors, typecheck 0, lint:use-server 0,
27/27 vitest on BottomTabBar.
